:pencil2: Fixed linting errors.

+++ app/core/config.py
@@ -24,7 +24,7 @@ DATABASE: Dict[str, Any] = {
 SECRET_KEY: str = os.getenv('SECRET_KEY', secrets.token_urlsafe(32))
 
 # ~~~~~ JWT ~~~~~
-ACCESS_TOKEN_EXPIRE_MINUTES: int = os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES', 60 * 24 * 8)
+ACCESS_TOKEN_EXPIRE_MINUTES: int = int(os.getenv('ACCESS_TOKEN_EXPIRE_MINUTES', 60 * 24 * 8))
 
 # ~~~~~ LOGGING ~~~~~
 LOGLEVEL: str = os.getenv('LOGLEVEL', 'INFO')
